Splitting HostController (not done)
