+++ server/src/main/java/umm3601/host/HostController.java
@@ -121,54 +121,54 @@ public class HostController implements Controller {
     }
   }
 
-  public Hunt getHunt(Context ctx) {
-    String id = ctx.pathParam("id");
-    Hunt hunt;
+  // public Hunt getHunt(Context ctx) {
+  //   String id = ctx.pathParam("id");
+  //   Hunt hunt;
 
-    try {
-      hunt = huntCollection.find(eq("_id", new ObjectId(id))).first();
-    } catch (IllegalArgumentException e) {
-      throw new BadRequestResponse("The requested hunt id wasn't a legal Mongo Object ID.");
-    }
-    if (hunt == null) {
-      throw new NotFoundResponse("The requested hunt was not found");
-    } else {
-      return hunt;
-    }
-  }
+  //   try {
+  //     hunt = huntCollection.find(eq("_id", new ObjectId(id))).first();
+  //   } catch (IllegalArgumentException e) {
+  //     throw new BadRequestResponse("The requested hunt id wasn't a legal Mongo Object ID.");
+  //   }
+  //   if (hunt == null) {
+  //     throw new NotFoundResponse("The requested hunt was not found");
+  //   } else {
+  //     return hunt;
+  //   }
+  // }
 
-  public void getHunts(Context ctx) {
-    Bson combinedFilter = constructFilterHunts(ctx);
-    Bson sortingOrder = constructSortingOrderHunts(ctx);
+  // public void getHunts(Context ctx) {
+  //   Bson combinedFilter = constructFilterHunts(ctx);
+  //   Bson sortingOrder = constructSortingOrderHunts(ctx);
 
-    ArrayList<Hunt> matchingHunts = huntCollection
-        .find(combinedFilter)
-        .sort(sortingOrder)
-        .into(new ArrayList<>());
+  //   ArrayList<Hunt> matchingHunts = huntCollection
+  //       .find(combinedFilter)
+  //       .sort(sortingOrder)
+  //       .into(new ArrayList<>());
 
-    ctx.json(matchingHunts);
+  //   ctx.json(matchingHunts);
 
-    ctx.status(HttpStatus.OK);
-  }
+  //   ctx.status(HttpStatus.OK);
+  // }
 
-  private Bson constructFilterHunts(Context ctx) {
-    List<Bson> filters = new ArrayList<>();
+  // private Bson constructFilterHunts(Context ctx) {
+  //   List<Bson> filters = new ArrayList<>();
 
-    if (ctx.queryParamMap().containsKey(HOST_KEY)) {
-      String targetHost = ctx.queryParamAsClass(HOST_KEY, String.class).get();
-      filters.add(eq(HOST_KEY, targetHost));
-    }
+  //   if (ctx.queryParamMap().containsKey(HOST_KEY)) {
+  //     String targetHost = ctx.queryParamAsClass(HOST_KEY, String.class).get();
+  //     filters.add(eq(HOST_KEY, targetHost));
+  //   }
 
-    Bson combinedFilter = filters.isEmpty() ? new Document() : and(filters);
+  //   Bson combinedFilter = filters.isEmpty() ? new Document() : and(filters);
 
-    return combinedFilter;
-  }
+  //   return combinedFilter;
+  // }
 
-  private Bson constructSortingOrderHunts(Context ctx) {
-    String sortBy = Objects.requireNonNullElse(ctx.queryParam("sortby"), "name");
-    Bson sortingOrder = Sorts.ascending(sortBy);
-    return sortingOrder;
-  }
+  // private Bson constructSortingOrderHunts(Context ctx) {
+  //   String sortBy = Objects.requireNonNullElse(ctx.queryParam("sortby"), "name");
+  //   Bson sortingOrder = Sorts.ascending(sortBy);
+  //   return sortingOrder;
+  // }
 
   public ArrayList<Task> getTasks(Context ctx) {
     Bson sortingOrder = constructSortingOrderTasks(ctx);
@@ -189,20 +189,20 @@ public class HostController implements Controller {
     return sortingOrder;
   }
 
-  public void addNewHunt(Context ctx) {
-    Hunt newHunt = ctx.bodyValidator(Hunt.class)
-    .check(hunt -> hunt.hostId != null && hunt.hostId.length() > 0, "Invalid hostId")
-    .check(hunt -> hunt.name.length() <= REASONABLE_NAME_LENGTH_HUNT, "Name must be less than 50 characters")
-    .check(hunt -> hunt.name.length() > 0, "Name must be at least 1 character")
-    .check(hunt -> hunt.description.length() <= REASONABLE_DESCRIPTION_LENGTH_HUNT,
-     "Description must be less than 200 characters")
-    .check(hunt -> hunt.est <= REASONABLE_EST_LENGTH_HUNT, "Estimated time must be less than 4 hours")
-    .get();
+  // public void addNewHunt(Context ctx) {
+  //   Hunt newHunt = ctx.bodyValidator(Hunt.class)
+  //   .check(hunt -> hunt.hostId != null && hunt.hostId.length() > 0, "Invalid hostId")
+  //   .check(hunt -> hunt.name.length() <= REASONABLE_NAME_LENGTH_HUNT, "Name must be less than 50 characters")
+  //   .check(hunt -> hunt.name.length() > 0, "Name must be at least 1 character")
+  //   .check(hunt -> hunt.description.length() <= REASONABLE_DESCRIPTION_LENGTH_HUNT,
+  //    "Description must be less than 200 characters")
+  //   .check(hunt -> hunt.est <= REASONABLE_EST_LENGTH_HUNT, "Estimated time must be less than 4 hours")
+  //   .get();
 
-    huntCollection.insertOne(newHunt);
-    ctx.json(Map.of("id", newHunt._id));
-    ctx.status(HttpStatus.CREATED);
-  }
+  //   huntCollection.insertOne(newHunt);
+  //   ctx.json(Map.of("id", newHunt._id));
+  //   ctx.status(HttpStatus.CREATED);
+  // }
 
   public void addNewTask(Context ctx) {
     Task newTask = ctx.bodyValidator(Task.class)
@@ -228,49 +228,49 @@ public class HostController implements Controller {
     }
   }
 
-  public void deleteHunt(Context ctx) {
-    String id = ctx.pathParam("id");
-    DeleteResult deleteResult = huntCollection.deleteOne(eq("_id", new ObjectId(id)));
-    if (deleteResult.getDeletedCount() != 1) {
-      ctx.status(HttpStatus.NOT_FOUND);
-      throw new NotFoundResponse(
-          "Was unable to delete ID "
-              + id
-              + "; perhaps illegal ID or an ID for an item not in the system?");
-    }
-    deleteTasks(ctx);
-    ctx.status(HttpStatus.OK);
-  }
+  // public void deleteHunt(Context ctx) {
+  //   String id = ctx.pathParam("id");
+  //   DeleteResult deleteResult = huntCollection.deleteOne(eq("_id", new ObjectId(id)));
+  //   if (deleteResult.getDeletedCount() != 1) {
+  //     ctx.status(HttpStatus.NOT_FOUND);
+  //     throw new NotFoundResponse(
+  //         "Was unable to delete ID "
+  //             + id
+  //             + "; perhaps illegal ID or an ID for an item not in the system?");
+  //   }
+  //   deleteTasks(ctx);
+  //   ctx.status(HttpStatus.OK);
+  // }
 
-  public void deleteTask(Context ctx) {
-    String id = ctx.pathParam("id");
-    try {
-      String huntId = taskCollection.find(eq("_id", new ObjectId(id))).first().huntId;
-      taskCollection.deleteOne(eq("_id", new ObjectId(id)));
-      decreaseTaskCount(huntId);
-    } catch (Exception e) {
-      ctx.status(HttpStatus.NOT_FOUND);
-      throw new NotFoundResponse(
-          "Was unable to delete ID "
-              + id
-              + "; perhaps illegal ID or an ID for an item not in the system?");
-    }
-    ctx.status(HttpStatus.OK);
-  }
+  // public void deleteTask(Context ctx) {
+  //   String id = ctx.pathParam("id");
+  //   try {
+  //     String huntId = taskCollection.find(eq("_id", new ObjectId(id))).first().huntId;
+  //     taskCollection.deleteOne(eq("_id", new ObjectId(id)));
+  //     decreaseTaskCount(huntId);
+  //   } catch (Exception e) {
+  //     ctx.status(HttpStatus.NOT_FOUND);
+  //     throw new NotFoundResponse(
+  //         "Was unable to delete ID "
+  //             + id
+  //             + "; perhaps illegal ID or an ID for an item not in the system?");
+  //   }
+  //   ctx.status(HttpStatus.OK);
+  // }
 
-  public void decreaseTaskCount(String huntId) {
-    try {
-      huntCollection.findOneAndUpdate(eq("_id", new ObjectId(huntId)),
-          new Document("$inc", new Document("numberOfTasks", -1)));
-    } catch (Exception e) {
-      e.printStackTrace();
-    }
-  }
+  // public void decreaseTaskCount(String huntId) {
+  //   try {
+  //     huntCollection.findOneAndUpdate(eq("_id", new ObjectId(huntId)),
+  //         new Document("$inc", new Document("numberOfTasks", -1)));
+  //   } catch (Exception e) {
+  //     e.printStackTrace();
+  //   }
+  // }
 
-  public void deleteTasks(Context ctx) {
-    String huntId = ctx.pathParam("id");
-    taskCollection.deleteMany(eq("huntId", huntId));
-  }
+  // public void deleteTasks(Context ctx) {
+  //   String huntId = ctx.pathParam("id");
+  //   taskCollection.deleteMany(eq("huntId", huntId));
+  // }
 
   public void getCompleteHunt(Context ctx) {
     CompleteHunt completeHunt = new CompleteHunt();
