+++ server/src/main/java/umm3601/host/HuntController.java
@@ -1,0 +1,181 @@
+package umm3601.host;
+
+import io.javalin.Javalin;
+import io.javalin.http.BadRequestResponse;
+import io.javalin.http.Context;
+import io.javalin.http.HttpStatus;
+import io.javalin.http.NotFoundResponse;
+import umm3601.Controller;
+
+import static com.mongodb.client.model.Filters.and;
+import static com.mongodb.client.model.Filters.eq;
+
+import java.io.File;
+import java.io.IOException;
+import java.io.InputStream;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.Paths;
+import java.nio.file.StandardCopyOption;
+import java.util.Date;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Map;
+import java.util.Objects;
+import java.util.Random;
+import java.util.UUID;
+
+import org.bson.Document;
+import org.bson.UuidRepresentation;
+import org.bson.conversions.Bson;
+import org.bson.types.ObjectId;
+import org.mongojack.JacksonMongoCollection;
+
+import com.mongodb.client.MongoDatabase;
+import com.mongodb.client.model.Sorts;
+import com.mongodb.client.result.DeleteResult;
+import java.util.Base64;
+
+public class HuntController implements Controller {
+
+  private static final String API_HUNT = "/api/hunts/{id}";
+  private static final String API_HUNTS = "/api/hunts";
+
+  static final String HOST_KEY = "hostId";
+  static final String HUNT_KEY = "huntId";
+
+  static final int REASONABLE_NAME_LENGTH_HUNT = 50;
+  static final int REASONABLE_DESCRIPTION_LENGTH_HUNT = 200;
+  private static final int  REASONABLE_EST_LENGTH_HUNT = 240;
+
+  private final JacksonMongoCollection<Hunt> huntCollection;
+  private final JacksonMongoCollection<Task> taskCollection;
+
+  public HuntController(MongoDatabase database) {
+    huntCollection = JacksonMongoCollection.builder().build(
+      database,
+      "hunts",
+      Hunt.class,
+      UuidRepresentation.STANDARD);
+
+    taskCollection = JacksonMongoCollection.builder().build(
+      database,
+      "tasks",
+      Task.class,
+      UuidRepresentation.STANDARD);
+
+
+    File directory = new File("photos");
+    if (!directory.exists()) {
+      directory.mkdir();
+    }
+  }
+
+  public Hunt getHunt(Context ctx) {
+    String id = ctx.pathParam("id");
+    Hunt hunt;
+
+    try {
+      hunt = huntCollection.find(eq("_id", new ObjectId(id))).first();
+    } catch (IllegalArgumentException e) {
+      throw new BadRequestResponse("The requested hunt id wasn't a legal Mongo Object ID.");
+    }
+    if (hunt == null) {
+      throw new NotFoundResponse("The requested hunt was not found");
+    } else {
+      return hunt;
+    }
+  }
+
+  public void getHunts(Context ctx) {
+    Bson combinedFilter = constructFilterHunts(ctx);
+    Bson sortingOrder = constructSortingOrderHunts(ctx);
+
+    ArrayList<Hunt> matchingHunts = huntCollection
+        .find(combinedFilter)
+        .sort(sortingOrder)
+        .into(new ArrayList<>());
+
+    ctx.json(matchingHunts);
+
+    ctx.status(HttpStatus.OK);
+  }
+
+  private Bson constructFilterHunts(Context ctx) {
+    List<Bson> filters = new ArrayList<>();
+
+    if (ctx.queryParamMap().containsKey(HOST_KEY)) {
+      String targetHost = ctx.queryParamAsClass(HOST_KEY, String.class).get();
+      filters.add(eq(HOST_KEY, targetHost));
+    }
+
+    Bson combinedFilter = filters.isEmpty() ? new Document() : and(filters);
+
+    return combinedFilter;
+  }
+
+  private Bson constructSortingOrderHunts(Context ctx) {
+    String sortBy = Objects.requireNonNullElse(ctx.queryParam("sortby"), "name");
+    Bson sortingOrder = Sorts.ascending(sortBy);
+    return sortingOrder;
+  }
+
+  public void addNewHunt(Context ctx) {
+    Hunt newHunt = ctx.bodyValidator(Hunt.class)
+    .check(hunt -> hunt.hostId != null && hunt.hostId.length() > 0, "Invalid hostId")
+    .check(hunt -> hunt.name.length() <= REASONABLE_NAME_LENGTH_HUNT, "Name must be less than 50 characters")
+    .check(hunt -> hunt.name.length() > 0, "Name must be at least 1 character")
+    .check(hunt -> hunt.description.length() <= REASONABLE_DESCRIPTION_LENGTH_HUNT,
+     "Description must be less than 200 characters")
+    .check(hunt -> hunt.est <= REASONABLE_EST_LENGTH_HUNT, "Estimated time must be less than 4 hours")
+    .get();
+
+    huntCollection.insertOne(newHunt);
+    ctx.json(Map.of("id", newHunt._id));
+    ctx.status(HttpStatus.CREATED);
+  }
+
+  public void deleteHunt(Context ctx) {
+    String id = ctx.pathParam("id");
+    DeleteResult deleteResult = huntCollection.deleteOne(eq("_id", new ObjectId(id)));
+    if (deleteResult.getDeletedCount() != 1) {
+      ctx.status(HttpStatus.NOT_FOUND);
+      throw new NotFoundResponse(
+          "Was unable to delete ID "
+              + id
+              + "; perhaps illegal ID or an ID for an item not in the system?");
+    }
+    deleteTasks(ctx);
+    ctx.status(HttpStatus.OK);
+  }
+
+  public void deleteTask(Context ctx) {
+    String id = ctx.pathParam("id");
+    try {
+      String huntId = taskCollection.find(eq("_id", new ObjectId(id))).first().huntId;
+      taskCollection.deleteOne(eq("_id", new ObjectId(id)));
+      decreaseTaskCount(huntId);
+    } catch (Exception e) {
+      ctx.status(HttpStatus.NOT_FOUND);
+      throw new NotFoundResponse(
+          "Was unable to delete ID "
+              + id
+              + "; perhaps illegal ID or an ID for an item not in the system?");
+    }
+    ctx.status(HttpStatus.OK);
+  }
+
+  public void decreaseTaskCount(String huntId) {
+    try {
+      huntCollection.findOneAndUpdate(eq("_id", new ObjectId(huntId)),
+          new Document("$inc", new Document("numberOfTasks", -1)));
+    } catch (Exception e) {
+      e.printStackTrace();
+    }
+  }
+
+  public void deleteTasks(Context ctx) {
+    String huntId = ctx.pathParam("id");
+    taskCollection.deleteMany(eq("huntId", huntId));
+  }
+}
